Conflicts:
	Assets/Project/Prefab/chest_epic.prefab
	Assets/Project/Script/Character.meta
	Assets/Project/Script/Manager/GameManager.cs
	Assets/Project/Script/Manager/LevelManager.cs
	Assets/Resources.meta
	ProjectSettings/TagManager.asset

+++ Assets/Project/Script/Gui/Bar/ManaBar.cs
@@ -6,7 +6,7 @@ public class ManaBar : Bar {
     void Update()
     {
         Characteristics player_stats = player.CharacterStats.UnitCharacteristics;
-        float mana_ratio = (float)player_stats.Mana / (float)player_stats.MaxMana;
+        float mana_ratio = player_stats.Mana / player_stats.MaxMana;
 
         if (player_stats.Mana >= 0)
         {
